requirements_txt.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.14

# YouTube processing
youtube-transcript-api==0.6.1
pytube==15.0.0
yt-dlp==2023.11.16

# Text processing and NLP
langchain==0.0.350
langchain-community==0.0.1
openai==1.3.7
tiktoken==0.5.2
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
pandas==2.1.4

# Vector database and embeddings
chromadb==0.4.18
faiss-cpu==1.7.4
pinecone-client==2.2.4

# Web scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
sqlite3  # Built-in with Python

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3
regex==2023.10.3
